Cleanup some integer type conversion warnings
* NSInteger is 'int' on 32-bit and 'long' on 64-bit, so there were
  some fun warnings about usage in formatting strings where %d was
  being used when building for 64-bit device
* Some warnings about enum types

+++ wikipedia/View Controllers/WebView/WebViewController.m
@@ -227,7 +227,7 @@ typedef enum {
 {
     NSDate *lastHousekeepingDate = [[NSUserDefaults standardUserDefaults] objectForKey:@"LastHousekeepingDate"];
     NSInteger daysSinceLastHouseKeeping = [[NSDate date] daysAfterDate:lastHousekeepingDate];
-    NSLog(@"daysSinceLastHouseKeeping = %d", daysSinceLastHouseKeeping);
+    NSLog(@"daysSinceLastHouseKeeping = %ld", (long)daysSinceLastHouseKeeping);
     if (daysSinceLastHouseKeeping > 1) {
         NSLog(@"Performing housekeeping...");
         CoreDataHousekeeping *imageHousekeeping = [[CoreDataHousekeeping alloc] init];
